Change implem removing js DOM manipulation
