+++ packages/theme/src/theme/_forms.scss
@@ -3,6 +3,9 @@
 /// @group Forms
 ////
 
+$form-shadow: inset 0 -1px 0 $silver-chalice;
+$form-focus-shadow: inset 0 -2px 0 $scooter;
+
 form {
 	// Bootswatch Paper theme fork
 	// @see https://github.com/thomaspark/bootswatch/tree/gh-pages/paper
@@ -49,7 +52,7 @@ form {
 	[type=email].form-control,
 	[type=tel].form-control,
 	[contenteditable].form-control {
-		@include box-shadow(inset 0 -1px 0 $silver-chalice);
+		@include box-shadow($form-shadow);
 		padding: 0;
 		border: none;
 		border-radius: 0;
@@ -57,7 +60,7 @@ form {
 		font-weight: $form-base-font-weight;
 
 		&:focus {
-			@include box-shadow(inset 0 -2px 0 $scooter);
+			@include box-shadow($form-focus-shadow);
 
 		}
 
@@ -89,7 +92,7 @@ form {
 
 	select,
 	select.form-control {
-		@include box-shadow(inset 0 -1px 0 $silver-chalice);
+		@include box-shadow($form-shadow);
 		border: none;
 		border-radius: 0;
 		appearance: none;
